+++ src/django_elasticsearch_dsl_drf/filter_backends/ordering/__init__.py
@@ -2,10 +2,15 @@
 Ordering backends.
 """
 
-from .common import *
-from .geo_spatial import *
+from .common import DefaultOrderingFilterBackend, OrderingFilterBackend
+from .geo_spatial import GeoSpatialOrderingFilterBackend
 
 __title__ = 'django_elasticsearch_dsl_drf.filter_backends.ordering'
 __author__ = 'Artur Barseghyan <artur.barseghyan@gmail.com>'
 __copyright__ = '2017 Artur Barseghyan'
 __license__ = 'GPL 2.0/LGPL 2.1'
+__all__ = (
+    'DefaultOrderingFilterBackend',
+    'GeoSpatialOrderingFilterBackend',
+    'OrderingFilterBackend',
+)